Move Document to its own file
